third_party/WebKit/LayoutTests/media/controls/controls-cast-button-low-end-device.html

<!DOCTYPE html>
<title>media controls cast button low end device</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<video width="500"></video>
<script>
async_test(function(t) {
    internals.setIsLowEndDevice(true);
    t.add_cleanup(_ => {
      internals.setIsLowEndDevice(false);
    });

    var video = document.querySelector("video");
    console.log("Adding video attribute controls");
    video.setAttribute("controls", "");
    console.log("Added video attribute controls");

    setTimeout(t.step_func(_ => {
      setTimeout(t.step_func_done(_ => {
        console.log("Getting the cast button");
        var button = castButton(video);
        console.log("Cast button display is " + button.style.display);
        assert_true(("display" in button.style) && (button.style.display != "none"), "button should exist");
      }), 0);
    }), 0);
});
</script>
